upgraded values in info to long long added really short example python dbpreader-using script removed unused schedulers.h
--HG--
branch : stealing_instrumentation

+++ src/pins/papi/cachemiss.h
@@ -13,7 +13,7 @@ typedef struct pins_cachemiss_info_s {
 	int kernel_type;
 	int th_id;
 	int values_len;
-	int values[NUM_EXEC_EVENTS];
+	long long values[NUM_EXEC_EVENTS];
 } pins_cachemiss_info_t;
 
 void pins_init_cachemiss(dague_context_t * master_context);
